Collapse LidSpaceCompactionJobBase into CompactionJob.
