+++ searchcore/src/vespa/searchcore/proton/server/CMakeLists.txt
@@ -57,7 +57,6 @@ vespa_add_library(searchcore_server STATIC
     ireplayconfig.cpp
     job_tracked_maintenance_job.cpp
     lid_space_compaction_handler.cpp
-    lid_space_compaction_job_base.cpp
     lid_space_compaction_job_take2.cpp
     maintenance_controller_explorer.cpp
     maintenance_jobs_injector.cpp
